+++ changelog.md
@@ -55,6 +55,7 @@
 + Fixed: incorrect row deletion when records table are update during confirmation
 + Fixed duplicate results in teamwork
 + Fixed: groups and organizations lost their sort order when sending selected items to a report template (HTML/DOCX/CSV)
++ Fixed: lost results in a multiday race when sending selected items to a report template (HTML/DOCX/CSV)
 
 ## 2025-04-01